Erstattet alle forekomster at AfxGetApp()-> med theApp. samt tilføjet erklæring af extern ...theApp i App....h-filerne. og tilføjet #include af af disse i stdafx.h mangesteder er SetWindowText erstattet af setWindowText(this,id,str)

+++ Test/TextPacker/TextPacker.cpp
@@ -14,7 +14,7 @@ WordSpace::WordSpace(const String &s) : String(s) {
 }
 
 int WordSpace::length() const {
-  return String::length() + m_spaceCount;
+  return (int)String::length() + m_spaceCount;
 }
 
 class TextPacker {
